server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "requests>=2.32.5",
]
